Update README.md
This Scratch project, "Jungle Jaunt" is an animation of a lion strolling through the jungle. Two butterflies are in the area as well - the small one glides around at will, but the viewer must click the mouse pointer on the larger butterfly to make it move. Every time the viewer clicks the larger butterfly, it will move to a new random spot.